separate views.py
