+++ app/views/__init__.py
@@ -1,0 +1,5 @@
+import app.views.searchlist_view
+import app.views.detail_view
+import app.views.register_view
+import app.views.update_view
+import app.views.delete_view
