+++ app/views/delete_view.py
@@ -1,0 +1,13 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.urls import reverse_lazy
+from django.views.generic.edit import DeleteView
+
+
+
+from app.models import Item
+
+
+# 削除画面
+class ItemDeleteView(LoginRequiredMixin, DeleteView):
+    model = Item
+    success_url = reverse_lazy('index')
