+++ app/views/searchlist_view.py
@@ -1,13 +1,9 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.urls import reverse_lazy
-from django.views.generic import DetailView
-from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django_filters.views import FilterView
 
-from .filters import ItemFilter
-from .forms import ItemForm
-from .models import Item
 
+from app.filters import ItemFilter
+from app.models import Item
 
 # Create your views here.
 # 検索一覧画面
@@ -35,28 +31,3 @@ class ItemFilterView(LoginRequiredMixin, FilterView):
                     request.GET[key] = request.session['query'][key]
 
         return super().get(request, **kwargs)
-
-
-# 詳細画面
-class ItemDetailView(LoginRequiredMixin, DetailView):
-    model = Item
-
-
-# 登録画面
-class ItemCreateView(LoginRequiredMixin, CreateView):
-    model = Item
-    form_class = ItemForm
-    success_url = reverse_lazy('index')
-
-
-# 更新画面
-class ItemUpdateView(LoginRequiredMixin, UpdateView):
-    model = Item
-    form_class = ItemForm
-    success_url = reverse_lazy('index')
-
-
-# 削除画面
-class ItemDeleteView(LoginRequiredMixin, DeleteView):
-    model = Item
-    success_url = reverse_lazy('index')
